Move most of bounded_value to constrained_value.
bounded_value was too tightly coupled with the idea of minimum and
maximum values, but the point of the validator classes was to maximum
flexibility (e.g., a validator that can ensure numbers are a power of
two).  constrained_value is much more generic, and bounded_value (and
clamped_value) are now type aliases to constrained_value.